# Store hashmap values in a single Vec instead of Rc's