--- a/benches/bench.rs
+++ b/benches/bench.rs
@@ -2,8 +2,8 @@
 
 extern crate test;
 
-use test::{Bencher, black_box};
 use bimap::BiHashMap;
+use test::{black_box, Bencher};
 
 #[bench]
 fn bench_hash_map_insert(b: &mut Bencher) {
@@ -12,6 +12,19 @@ fn bench_hash_map_insert(b: &mut Bencher) {
 
         for i in 1i64..100 {
             map.insert(format!("key {}", i), i);
+        }
+
+        black_box(map);
+    });
+}
+
+#[bench]
+fn bench_hash_map_insert_no_overwrite(b: &mut Bencher) {
+    b.iter(|| {
+        let mut map: BiHashMap<String, i64> = BiHashMap::new();
+
+        for i in 1i64..100 {
+            let _ = map.insert_no_overwrite(format!("key {}", i), i);
         }
 
         black_box(map);