This stores all the value pairs in a Vec and keeps 2 HashMaps that hold indexes into the Vec to allow lookups.

This trades the overhead from Rc with the overhead of using the index, note that in both cases the HashMaps hold "pointers" to the values.

On my local machine the added benchmarks give the following result

before:

    test bench_hash_map_get_left            ... bench:       8,294 ns/iter (+/- 724)
    test bench_hash_map_get_right           ... bench:       2,094 ns/iter (+/- 286)
    test bench_hash_map_insert              ... bench:      34,270 ns/iter (+/- 1,634)
    test bench_hash_map_insert_no_overwrite ... bench:      33,936 ns/iter (+/- 4,649)
    test bench_hash_map_iter                ... bench:         320 ns/iter (+/- 47)
    test bench_hash_map_replace_both        ... bench:      27,046 ns/iter (+/- 2,294)
    test bench_hash_map_replace_left        ... bench:      35,141 ns/iter (+/- 2,200)
    test bench_hash_map_replace_right       ... bench:      31,532 ns/iter (+/- 2,925)
    test bench_hash_map_replace_same        ... bench:      23,992 ns/iter (+/- 1,224)

after:

    test bench_hash_map_get_left            ... bench:       7,807 ns/iter (+/- 1,015)
    test bench_hash_map_get_right           ... bench:       2,195 ns/iter (+/- 321)
    test bench_hash_map_insert              ... bench:      16,577 ns/iter (+/- 1,068)
    test bench_hash_map_insert_no_overwrite ... bench:      15,947 ns/iter (+/- 1,006)
    test bench_hash_map_iter                ... bench:          66 ns/iter (+/- 7)
    test bench_hash_map_replace_both        ... bench:      16,406 ns/iter (+/- 1,372)
    test bench_hash_map_replace_left        ... bench:      21,219 ns/iter (+/- 2,122)
    test bench_hash_map_replace_right       ... bench:      18,556 ns/iter (+/- 1,374)
    test bench_hash_map_replace_same        ... bench:      11,585 ns/iter (+/- 1,510)

Getting values is pretty much the same since both have one layer of indirection on top of the HashMap.
Inserting is ~30% faster, presumably due to not having to create the Rc's.
Iterating is a lot faster presumably due the cache locality of the values in the new setup.

Memory overhead should be less -- the Rc adds 2 usize's overhead per stored item for refcounting, both implementations store 2 "pointers" per value in the HashMaps) -- but I don't have proper numbers on that.

Since the keys stored in the new HashMaps are themselves already hashes of the inserted values, a dummy hasher is used to prevent double hashing.