@billyrieger 2019-06-08
Sorry for the slow reply, my computer was out of commission for a while. This seems like it would be a good improvement for the library. Before going too deep into the code, I have a question regarding the `NullHasher`. Instead of having a `HashMap<u64, usize, NullHasherBuilder>` and hashing the keys manually with `Self::get_hash()`, wouldn't it make more sense just to have a `HashMap<T, usize>` to store the data indices? How it's written now seems to have an unnecessary layer of indirection (hence why `NullHasher` is needed).