@billyrieger 2019-06-08
Also, if we're doing this change for `BiHashMap` it makes sense to also do it for `BiBTreeMap`.