@icewind1991 2019-06-09
using a `HashMap<T, usize>` would require storing the values in the hashmap, by using the hash we can put them in the `Vec` instead.

Updating `BiBTreeMap` is generally something I want to do once this is merged, and I also want to see if I can generalize the bimap to work with any underlying map