@icewind1991 2019-06-10
I just realized that doing the pre-hashing removes the hashmap's ability to handle hash collisions and will thus not be reliable.

I'll see if I can adjust the way the data is stored