@billyrieger 2020-07-26
I'm closing this for now, but if you ever get something to work you can open a new pull request.